refactor: refactor form
