--- src/components/Form/index.ts
@@ -1,14 +1,0 @@
-'use client';
-
-export type { TFormProps, TFormConfigType, TValidationProps, TAsterisk } from './typing';
-export { Form } from './Form';
-export { ValidationTypes, ValidationPattern, FormFieldNames } from './constants';
-export {
-  exampleConfig,
-  signinConfig,
-  signupConfig,
-  profileConfig,
-  bookingConfig,
-  callbackConfig,
-} from './configs';
-export { StFieldset, StFieldWrapper, StAsterick, StRelativeBox } from './styled';
