+++ src/components/UI/Checkbox/typing.ts
@@ -5,7 +5,6 @@ import { UseFormRegisterReturn, UseFormWatch } from 'react-hook-form';
 export type TCheckbox = {
   name: FormFieldNames;
   register: UseFormRegisterReturn<FormFieldNames>;
-  required: boolean;
   pattern?: ValidationTypes;
   watch: UseFormWatch<Record<string, string>>;
   label?: string;
