+++ src/features/Form/index.ts
@@ -1,0 +1,5 @@
+'use client';
+
+export { Form } from './ui/Form/Form';
+export type { TFormConfigType } from './typing';
+export { ValidationTypes, FormFieldNames } from './consts';
