+++ src/features/Form/ui/Field/styled.ts
@@ -1,19 +1,17 @@
 import styled from 'styled-components';
 import { StFlex } from '@/styles/global';
-import { TAsterisk } from './';
+import type { TAsterisk } from '../../typing';
 import { theme } from '@/styles/theme';
-
-export const StFieldset = styled.fieldset`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
 
 export const StFieldWrapper = styled(StFlex)`
   position: relative;
   width: 346px;
   flex-direction: column;
   gap: 8px;
+`;
+
+export const StRelativeBox = styled.div`
+  position: relative;
 `;
 
 export const StAsterick = styled.div<TAsterisk>`
@@ -25,8 +23,4 @@ export const StAsterick = styled.div<TAsterisk>`
   left: -17px;
   color: ${theme.colors.error};
   visibility: ${({ visibility }) => visibility && visibility};
-`;
-
-export const StRelativeBox = styled.div`
-  position: relative;
 `;
