+++ src/features/Form/ui/Form/styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StFieldset = styled.fieldset`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+`;
